refactor: extract signals list redesign, trace-picker, and manage-signal-sheet restructure
Reverts signals list page, manage-signal-sheet folder restructure,
signal-sparkline, empty-row, trace-picker, and debugger-sessions
sidebar trace-picker usage back to dev versions to reduce PR scope.
These changes will be stacked in a follow-up PR.
