--- frontend/components/traces/trace-picker/index.tsx
@@ -1,16 +1,0 @@
-"use client";
-
-import { DataTableStateProvider } from "@/components/ui/infinite-datatable/model/datatable-store";
-
-import { FETCH_SIZE, tracePickerColumnOrder } from "./columns";
-import TracePickerContent, { type TracePickerProps } from "./trace-picker-content";
-
-export type { TracePickerProps };
-
-const TracePicker = (props: TracePickerProps) => (
-  <DataTableStateProvider defaultColumnOrder={tracePickerColumnOrder} pageSize={FETCH_SIZE}>
-    <TracePickerContent {...props} />
-  </DataTableStateProvider>
-);
-
-export default TracePicker;
